fix: set_exclbind() passes bool (1 byte) instead of c_int (4 bytes)
setsockopt expects sizeof(int) for SO_EXCLBIND, but the generic
setsockopt wrapper uses size_of::<T>() which is 1 for bool. The kernel
returns EINVAL because the option length is too short, making
set_exclbind() always fail.

Every other boolean socket option in the file correctly casts to c_int
first (e.g., keepalive as c_int, nodelay as c_int). Apply the same
pattern here.

Affects Solaris and illumos.

+++ library/std/src/sys/net/connection/socket/unix.rs
@@ -535,7 +535,7 @@ impl Socket {
     pub fn set_exclbind(&self, excl: bool) -> io::Result<()> {
         // not yet on libc crate
         const SO_EXCLBIND: i32 = 0x1015;
-        unsafe { setsockopt(self, libc::SOL_SOCKET, SO_EXCLBIND, excl) }
+        unsafe { setsockopt(self, libc::SOL_SOCKET, SO_EXCLBIND, excl as c_int) }
     }
 
     #[cfg(any(target_os = "solaris", target_os = "illumos"))]
